meta-openxt-base: fixes for building under Debian Trixie
Enable Dunfell OpenXT base builds on Debian Trixie / Devuan 6-class hosts:

- perl: restore host miniperl filesystem defines in xconfig.h after
  GCC 14 configure probes drop readdir/symlink/link
- unzip: drop bogus -DNO_DIR when host dirent is present
- strace-native: rename private sched_attr to avoid host header clash
- efibootmgr: fetch from rhboot main (rhinstaller/master is gone)
- grub-xen: resolve OE_CORE_PATH for TOPDIR/../layers layouts
- ovmf: make Intel e1000/PREBOOT embedding optional via PACKAGECONFIG

These changes are specifically to enable building under Debian Trixie.

+++ meta-openxt-base/recipes-core/ovmf/ovmf_git.bbappend
@@ -3,7 +3,6 @@ FILESEXTRAPATHS_prepend := "${THISDIR}/${PN}:"
 # xcp-ng patches taken from:
 # https://github.com/xcp-ng-rpms/edk2/tree/master/SOURCES
 SRC_URI += " \
-    https://downloadmirror.intel.com/29334/eng/PREBOOT.EXE;unpack=0;name=PREBOOT \
     file://xcp-ng-set-default-resolution-1024-768.patch \
     file://xcp-ng-add-xen-variable.patch \
     file://xcp-ng-add-xen-platform-device-id.patch \
@@ -11,17 +10,28 @@ SRC_URI += " \
     file://xcp-ng-keep-caching-enabled.patch \
 "
 
-DEPENDS_append += " \
-    unzip-native \
-"
+# Optional Intel E1000 UEFI NIC driver embedding (needs PREBOOT.EXE).
+# Intel's downloadmirror often returns 403 to automated fetchers; enable when
+# you have PREBOOT.EXE in DL_DIR (or a reachable SRC_URI mirror).
+PACKAGECONFIG[e1000] = ",,,"
+# Keep historical default for product builds; local.conf may remove it.
+PACKAGECONFIG_append_class-target = " e1000"
+
+SRC_URI += "${@bb.utils.contains('PACKAGECONFIG', 'e1000', \
+    'https://downloadmirror.intel.com/29334/eng/PREBOOT.EXE;unpack=0;name=PREBOOT', \
+    '', d)}"
+
+DEPENDS_append += "${@bb.utils.contains('PACKAGECONFIG', 'e1000', 'unzip-native', '', d)}"
 
 # PREBOOT.EXE, OS independent, version 24.3.
 SRC_URI[PREBOOT.md5sum] = "8660641e184dafdeb78b8ca1fbd837f7"
 SRC_URI[PREBOOT.sha256sum] = "83dac749d74a6a54d7451bee79f9e1d605c4e2775d6b524d39030b248989092a"
 
 do_extract_bootutil() {
-    mkdir -p "${S}/Intel3.5/EFIX64"
-    unzip -q -p "${WORKDIR}/PREBOOT.EXE" "APPS/EFI/EFIx64/E3522X2.EFI" > "${S}/Intel3.5/EFIX64/E3522X2.EFI"
+    if ${@bb.utils.contains('PACKAGECONFIG', 'e1000', 'true', 'false', d)}; then
+        mkdir -p "${S}/Intel3.5/EFIX64"
+        unzip -q -p "${WORKDIR}/PREBOOT.EXE" "APPS/EFI/EFIx64/E3522X2.EFI" > "${S}/Intel3.5/EFIX64/E3522X2.EFI"
+    fi
 }
 addtask do_extract_bootutil before do_configure after do_unpack
 do_extract_bootutil[doc] = "Extract Intel's proprietary E1000 NIC driver to be embedded in OVMF image."
@@ -29,17 +39,26 @@ do_extract_bootutil[depends] = "${PN}:do_prepare_recipe_sysroot"
 do_extract_bootutil[dirs] = "${B}"
 
 do_compile_class-target_append() {
-    bbnote "Building with E1000 (support for netboot)."
-    rm -rf ${S}/Build/Ovmf$OVMF_DIR_SUFFIX
-    ${S}/OvmfPkg/build.sh $PARALLEL_JOBS -a $OVMF_ARCH -b RELEASE -t ${FIXED_GCCVER} -D E1000_ENABLE -D XEN_VARIABLE_ENABLE=TRUE -D SECURE_BOOT_ENABLE=TRUE
-    ln ${build_dir}/FV/OVMF.fd ${WORKDIR}/ovmf/ovmf.e1000.fd
-    ln ${build_dir}/FV/OVMF_CODE.fd ${WORKDIR}/ovmf/ovmf.e1000.code.fd
+    if ${@bb.utils.contains('PACKAGECONFIG', 'e1000', 'true', 'false', d)}; then
+        bbnote "Building with E1000 (support for netboot)."
+        rm -rf ${S}/Build/Ovmf$OVMF_DIR_SUFFIX
+        ${S}/OvmfPkg/build.sh $PARALLEL_JOBS -a $OVMF_ARCH -b RELEASE -t ${FIXED_GCCVER} -D E1000_ENABLE -D XEN_VARIABLE_ENABLE=TRUE -D SECURE_BOOT_ENABLE=TRUE
+        ln ${build_dir}/FV/OVMF.fd ${WORKDIR}/ovmf/ovmf.e1000.fd
+        ln ${build_dir}/FV/OVMF_CODE.fd ${WORKDIR}/ovmf/ovmf.e1000.code.fd
+    fi
 }
 
 do_install_class-target_append() {
     install -d ${D}${datadir}/firmware
-    install -m 0600 ${WORKDIR}/ovmf/ovmf.e1000.fd ${D}${datadir}/firmware/ovmf.e1000.bin
-    ln -sf ovmf.e1000.bin ${D}${datadir}/firmware/ovmf.bin
+    if ${@bb.utils.contains('PACKAGECONFIG', 'e1000', 'true', 'false', d)}; then
+        install -m 0600 ${WORKDIR}/ovmf/ovmf.e1000.fd ${D}${datadir}/firmware/ovmf.e1000.bin
+        ln -sf ovmf.e1000.bin ${D}${datadir}/firmware/ovmf.bin
+    else
+        # Stock OVMF FD as firmware when E1000 embedding is disabled.
+        if [ -f ${WORKDIR}/ovmf/ovmf.fd ]; then
+            install -m 0600 ${WORKDIR}/ovmf/ovmf.fd ${D}${datadir}/firmware/ovmf.bin
+        fi
+    fi
 }
 
 PACKAGES += " \
